ENG-38: Fix swap bugs from refactor
